% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{berechne_mixtox}
\alias{berechne_mixtox}
\title{Mischtoxizitäten berechnen}
\usage{
berechne_mixtox(rq_data, optin_mischtox_S = FALSE)
}
\arguments{
\item{rq_data}{Tibble mit Daten der Funktion berechne_rq_ue()}
}
\value{
Tibble mit Mischtoxizitäten
}
\description{
Mischtoxizitäten berechnen
}
