% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{berechne_rq_ue}
\alias{berechne_rq_ue}
\title{Risikoquotienten (RQ) für Einzelstoffe berechnen}
\usage{
berechne_rq_ue(
  mv_daten,
  regulierungen = NULL,
  kriterien = NULL,
  robust3 = TRUE
)
}
\arguments{
\item{mv_daten}{Normalisierte MV-Daten}

\item{regulierungen}{Regulierungsdaten (dataframe)}

\item{kriterien}{Qualitätskriterien (dataframe)}

\item{robust3}{Logisch (Vorgabe: \code{TRUE}). Warnen, wenn QK mit Robustheit 3 überschritten sind (nicht in Daten enthalten)?}
}
\value{
Dataframe mit MV-Daten, ergänzt mit Risikoquotienten
}
\description{
Risikoquotienten (RQ) für Einzelstoffe berechnen
}
