% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{einheiten_normalisieren}
\alias{einheiten_normalisieren}
\title{Einheiten von MV-Daten normalisieren}
\usage{
einheiten_normalisieren(mvdata, wert, einheit, zieleinheit = "µg/l")
}
\arguments{
\item{mvdata}{MV-Daten entsprechend der Spezifikationen.}

\item{wert}{Name der Spalte, die den Messwert enthält.}

\item{einheit}{Name der Spalte, die die Einheit enthält.}

\item{zieleinheit}{Zieleinheit: Eine von g/l, mg/l, µ/l, ng/l, pg/l (Vorgabe: µg/l). Achtung: Einheit in Qualitätskriterien-Tabelle beachten.}
}
\value{
Dataframe mit den normalisierten MV-Daten
}
\description{
Generische Funktion, um die Einheiten von MV-Daten zu normalisieren, d.h. für jede Messung dieselbe Einheit einzustellen.
}
