% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{einlesen_bafu_lookup}
\alias{einlesen_bafu_lookup}
\title{BAFU-Namens-Lookup einlesen}
\usage{
einlesen_bafu_lookup(bafu_lookup_pfad, alle_felder = FALSE)
}
\arguments{
\item{bafu_lookup_pfad}{Pfad zur Exceldatei mit dem Namens-Lookup}

\item{alle_felder}{Logisch (Vorgabe: \code{FALSE}). Sollen alle Spalten der Tabelle exportiert werden?}
}
\value{
Dataframe mit Zuordnung der französischen und deutschen Bezeichnungen zur eindeutigen BAFU Parameter-ID
}
\description{
BAFU-Namens-Lookup einlesen
}
