% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{einlesen_mv_gbl}
\alias{einlesen_mv_gbl}
\title{MV-Daten des GBL (Kt. Bern) einlesen}
\usage{
einlesen_mv_gbl(
  mv_daten_pfad,
  vsa_lookup_pfad = NULL,
  bafu_lookup_pfad = NULL,
  bSaP = FALSE
)
}
\arguments{
\item{mv_daten_pfad}{Pfad zur kommaseparierten Datei mit den MV-Daten}

\item{vsa_lookup_pfad}{Pfad zur VSA-Tabelle \code{Tab_Substanzen.xlsx}}

\item{bafu_lookup_pfad}{Pfad zum BAFU-Namen-Lookup}

\item{bSaP}{Logisch (Vorgabe \code{FALSE}). Sollen berechnete Mischproben verwendet werden?}
}
\value{
Dataframe mit MV-Daten
}
\description{
Liest MV-Daten des GBL (Kt. Bern) ein und bereitet sie für die weitere Verarbeitung auf. Achtung: Falls \code{bSaP == TRUE}, werden 3.5-Tage-Mischproben, die im Intervall (Beginn- bis Enddatum ohne Zeit) der berechneten 14-Tage-Mischproben liegen, entfernt! Für die Auswertung akuter Überschreitungen muss deshalb \code{bSaP == FALSE} gesetzt werden.
}
\examples{
mv_daten_pfad <- system.file("extdata", "Daten_MV_GBL_2019_2020.txt", package = "mvwizr")

# Ab mvwizr v1.2 können vsa_lookup_pfad und bafu_lookup_pfad leer gelassen werden
# - es wird dann die mit dem Paket gebundelte Datei verwendet.
mvdaten <- einlesen_mv_gbl(mv_daten_pfad)

}
