% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{einlesen_nawa}
\alias{einlesen_nawa}
\title{NAWA-MV-Daten einlesen}
\usage{
einlesen_nawa(
  nawa_mv,
  vsa_lookup_pfad = NULL,
  bafu_lookup_pfad = NULL,
  delimiter = NA_character_,
  encoding = NA_character_,
  lang = NA_character_,
  parameter = NA_character_,
  header = NA_integer_
)
}
\arguments{
\item{nawa_mv}{Entweder Pfad zur Datei im NAWA-Format (Excel oder Text) oder ein Dataframe im NAWA-Format (z.B. WISKI-Export)}

\item{vsa_lookup_pfad}{Pfad zur VSA-Tabelle \code{Tab_Substanzen.xlsx}}

\item{bafu_lookup_pfad}{Pfad zum BAFU-Namen-Lookup}

\item{delimiter}{Trennzeichen für Textdateien. Falls \code{NULL}, wird versucht, das Trennzeichen zu erraten. Standard ist \code{NULL}.}

\item{encoding}{Encoding der Textdatei. Falls \code{NULL}, wird versucht, das Encoding zu erraten. Standard ist \code{NULL}.}

\item{lang}{Sprache der Datei. Falls \code{NULL}, wird versucht, die Sprache zu erraten. Mögliche Werte sind "DE" (Deutsch) oder "FR" (Französisch). Standard ist \code{NULL}.}

\item{parameter}{Angabe, was im "Parameter"-Feld der MV-Datei steht (da nicht konsistent verwendet). Mögliche Werte sind "BAFU_Parameter_ID" (für den BAFU-Schlüssel ohne Sonderzeichen), "BAFU_Bez_DE" (Deutsche Bezeichnung) oder "BAFU_Bez_FR" (Französische Bezeichnung). Falls \code{NULL}, wird versucht, das Parameter-Feld zu erraten. Standard ist \code{NULL}.}

\item{header}{Zeilennummer, ab der die eigentliche Tabelle beginnt. Falls \code{NULL}, wird versucht, die Header-Zeile zu erraten. Standard ist \code{NULL}.}
}
\value{
Dataframe mit MV-Daten
}
\description{
Liest MV-Daten im BAFU-NAWA-Format ein. Falls die Datei einen Header aufweist vor dem Start der eigentlichen Tabelle, so wird dieser übersprungen. Um mehrere Dateien im NAWA-Format auf einmal einzulesen, kann die Funktion \code{batch_einlesen_nawa()} verwendet werden.
}
\details{
Die Funktion liest entweder eine (einzelne) Datei im Excel (.xlsx)-Format ein oder als Text mit Trennzeichen (.csv, .txt). Die Funktion kann anstelle einer Datei auch ein Dataframe im NAWA-Format entgegennehmen (z.B. über Datenbank-Verbindung abgerufen).

Duplikate (z.B. auch für Messungen auf mehreren Geräten) werden entfernt, indem jeweils der höchste Messwert für die gleiche Probe (Station, Beginn- und Enddatum) und Substanz verwendet wird.

Bestimmungsgrenzen werden aus den eingelesenen Daten bestimmt, wobei die maximale und minimale Bestimmungsgrenze (über die gesamte Messperiode in den Daten) für jede Substanz ermittelt wird - die min. und max. Bestimmungsgrenzen sind also konstant im Datensatz. Falls eine Substanz keine Bestimmungsgrenze hat, so wird sie nicht entfernt, sondern es wird eine Warnung ausgegeben.

Alle Einheiten werden (für die weiteren Berechnungen) automatisch auf µg/l normalisiert.

Messungen von Substanzen, die nicht einer BAFU-Parameter-ID und einer VSA Substanz-ID zugeordnet werden können, werden entfernt.
}
\examples{
# Testen der Einlesefunktion mit Heuristik
nawa_mv_fr <- system.file("extdata", "NAWA_Bsp_4.xlsx", package = "mvwizr")

out <- einlesen_nawa(nawa_mv_fr)

# Testen der Einlesefunktion mit Angabe von Argumenten
out2 <- einlesen_nawa(nawa_mv_fr,
  lang = "FR", parameter = "BAFU_Bez_FR",
  header = 8L
)

# Vergleich der Resultate
identical(out, out2)

}
