% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{einlesen_regulierungen}
\alias{einlesen_regulierungen}
\title{Regulierungsdaten einlesen}
\usage{
einlesen_regulierungen(regulierungen_pfad)
}
\arguments{
\item{regulierungen_pfad}{Pfad zur Exceldatei mit den Regulierungsdaten}
}
\value{
Dataframe mit Regulierungsinformationen
}
\description{
Regulierungsdaten einlesen
}
