% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{einlesen_vsa_lookup}
\alias{einlesen_vsa_lookup}
\title{VSA ID_Substanz Lookup einlesen}
\usage{
einlesen_vsa_lookup(vsa_lookup_pfad, alle_felder = FALSE)
}
\arguments{
\item{vsa_lookup_pfad}{Pfad zur Excel-Datei mit dem ID_Substanz-Lookup des VSA}

\item{alle_felder}{Alle Spalten exportieren? Standardmässig werden nur die im Weiteren benötigten Spalten exportiert.}
}
\value{
Dataframe mit Lookups
}
\description{
VSA ID_Substanz Lookup einlesen
}
