% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{entferne_duplikate}
\alias{entferne_duplikate}
\title{Entferne Duplikate aus MV-Daten}
\usage{
entferne_duplikate(mv_data, dateiname = NULL)
}
\arguments{
\item{mv_data}{Dataframe mit MV-Daten, die Duplikate enthalten können.}

\item{dateiname}{Optionaler Dateiname, der in der Warnung ausgegeben wird, falls Duplikate gefunden werden (da Warnungen gesammelt am Schluss einer Skript-Ausführung erscheinen). Falls \code{NULL}, wird "MV-Daten" verwendet.}
}
\value{
Dataframe mit MV-Daten ohne Duplikate.
}
\description{
Entfernt Duplikate aus MV-Daten, indem jeweils der Datensatz mit dem höchsten Messwert für die gleiche Probe (Station, Beginn- und Enddatum) und Substanz verwendet wird. Duplikate werden anhand der Spalten \code{CODE}, \code{PROBEARTID}, \code{BEGINNPROBENAHME}, \code{ENDEPROBENAHME} und \code{PARAMETER} identifiziert.
}
