% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kriterien_mvwizr}
\alias{kriterien_mvwizr}
\title{Qualitätskriterien}
\format{
\subsection{\code{kriterien_mvwizr}}{

Eine Tabelle mit 10 Variablen und 441 Einträgen. Untenstehend sind alle Variablen mit ihren Datentypen beschrieben:
\describe{
\item{ID_Substanz}{VSA-ID der Substanz (integer)}
\item{P_chron}{Gibt an, ob CQK zu berücksichtigen für Mischtoxizität bei Pflanzen (integer)}
\item{I_chron}{Gibt an, ob CQK zu berücksichtigen für Mischtoxizität bei Invertebraten (integer)}
\item{V_chron}{Gibt an, ob CQK zu berücksichtigen für Mischtoxizität bei Vertebraten (integer)}
\item{P_akut}{Gibt an, ob AQK zu berücksichtigen für Mischtoxizität bei Pflanzen (integer)}
\item{I_akut}{Gibt an, ob AQK zu berücksichtigen für Mischtoxizität bei Invertebraten (integer)}
\item{V_akut}{Gibt an, ob AQK zu berücksichtigen für Mischtoxizität bei Vertebraten (integer)}
\item{Robustheit QK}{Robustheit des QK. Wert 1-3, wobei QK mit einem Wert von 3 nicht verwendet werden (integer)}
\item{CQK}{Chronisches Qualitätskriterium in µg/l (numeric)}
\item{AQK}{Akutes Qualitätskriterium in µg/l (numeric)}
}
}
}
\source{
VSA
}
\usage{
kriterien_mvwizr
}
\description{
Dieser Datensatz beinhaltet Qualitätskriterien des VSA zu Mikroverunreinigungen (ohne Duplikate).
}
\keyword{datasets}
