% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mvdaten_beispiel_mvwizr}
\alias{mvdaten_beispiel_mvwizr}
\title{MV-Beispieldaten des Kantons Bern}
\format{
\subsection{\code{mvdaten_beispiel_mvwizr}}{

Eine Tabelle bestehend aus 22 Variablen ca. 100k Beobachtungen/Messungen. Untenstehend sind alle Variablen mit ihren Datentypen beschrieben:
\describe{
\item{UID}{Eindeutiger Bezeichner für Messung (integer)}
\item{CODE}{Stationscode (character)}
\item{STANDORT}{Standort Station (character)}
\item{NAME}{Name des Gewässers (character)}
\item{PROBEARTID}{Art der Probe: Stichprobe (S), Mischprobe (SaP) oder berechnete Mischprobe (bSaP) (character)}
\item{BEGINNPROBENAHME}{Beginn der Probenahme (POSIXct; Locale Europe/Zurich)}
\item{ENDEPROBENAHME}{Ende der Probenahme (SaP) resp. leer (S) (POSIXct; Locale Europe/Zurich)}
\item{ID_Substanz}{Gematchte VSA-ID der Substanz (integer)}
\item{PARAMETERID_BAFU}{BAFU-Schlüssel für Substanz (character)}
\item{OPERATOR}{Falls vorhanden (<) liegt Wert unter Bestimmungsgrenze (BG) (character)}
\item{WERT_NUM}{Konzentrationswert in µg/l. Falls 0, liegt Wert unter BG (numeric)}
\item{Konz_inkl_BG}{Ursprünglicher Konzentrationswert zur Kontrolle, ggf. unter BG (numeric)}
\item{EINHEIT}{Einheit der Messung. Hier immer µg/l normalisiert (character)}
\item{MSTLTYP}{Messstellentyp (character)}
\item{PARAMETERGRUPPEID}{Parametergruppencode des GBL (integer)}
\item{PARAMETERGRUPPE}{Bezeichnung der Parametergruppe des GBL (character)}
\item{PARAMETER}{Stoffbezeichnung des GBL intern (character)}
\item{BEZEICHNUNG_BAFU}{Deutsche BAFU-Bezeichnung aus GBL Datenbank (character)}
\item{BAFU_Bez_DE}{Deutsche Bezeichnung gemäss Matching mit BAFU Lookup (character)}
\item{BAFU_Bez_FR}{Französische Bezeichnung gemäss Matching mit BAFU Lookup (character)}
\item{BG_max}{Maximale festgestellte Bestimmungsgrenze für Substanz aus Daten in µg/l (numeric)}
\item{BG_min}{Minimale festgestellte Bestimmungsgrenze für Substanz aus Daten in µg/l (numeric)}
}
}
}
\source{
GBL Kt. Bern
}
\usage{
mvdaten_beispiel_mvwizr
}
\description{
MV-Daten bestehend aus mehreren Datensätzen von unterschiedlichen Stationen für die Jahre 2019 und 2020.
Beinhaltet Stichproben, 3.5-Tage-Proben und Proben mit längerem Sampling-Intervall.
Folgende wichtige Punkte sollten beachtet werden:
\itemize{
\item Die Einheiten wurden alle auf µg/l normalisiert
\item Duplikate (unterschiedliche Werte für dieselbe Probe und Substanz) wurden (konservativ) entfernt
\item Einträge mit fehlender Substanz_ID oder PARAMETERID_BAFU wurden entfernt.
}
}
\keyword{datasets}
