% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_misch_mixtox_verlauf}
\alias{plot_misch_mixtox_verlauf}
\title{Verlauf von Mischungstoxizitäten plotten}
\usage{
plot_misch_mixtox_verlauf(
  rq_ue_daten,
  stationscode = NULL,
  jahr = NULL,
  modus = "andauernd",
  plot_zusammenfassung = "keine"
)
}
\arguments{
\item{rq_ue_daten}{Dataframe mit Output der Funktion \code{berechne_rq_ue()}}

\item{stationscode}{Station, für welche der Plot erstellt werden soll. Falls \code{NULL}, wird für alle Stationen geplottet (Facets)}

\item{jahr}{Jahr, für welches der Überschreitungsplot erstellt werden soll. Falls \code{NULL}, werden alle Jahre geplottet.}

\item{modus}{\itemize{
\item \code{"andauernd"}: Berücksichtigt nur Stoffe mit einem spezifischen andauernden Grenzwert in der GSchV. Nur Proben mit einer Dauer von >= 10 Tagen werden verwendet und unter Verwendung des CQK beurteilt.
\item \code{"kurzzeitig"}: Berücksichtigt nur Stoffe mit einem spezifischen akuten Grenzwert in der GSchV. Berücksichtigt alle Proben (ausser Stichproben) und verwendet das AQK zur Beurteilung.
}}

\item{plot_zusammenfassung}{Entweder \code{NULL} (Vorgabe), "stichproben" oder "mischproben". Falls einer der letzeren beiden, wird pro Jahr und nicht pro Monat aggregiert (entweder nur für Stichproben oder nur für Mischproben). Für Stichproben muss der \code{modus} auf "kurzzeitig" gesetzt werden.}
}
\value{
ggplot2 Plot-Objekt
}
\description{
Verlauf von Mischungstoxizitäten plotten
}
\examples{

# Ausführlicher Verlauf für andauernde Belastungen
plot_misch_mixtox_verlauf(rq_ue_beispiel_mvwizr, modus = "andauernd")

# Ausführlicher Verlauf für kurzzeitige Belastungen
plot_misch_mixtox_verlauf(rq_ue_beispiel_mvwizr, modus = "kurzzeitig")

# Zusammengefasster Verlauf für andauernde Belastungen
plot_misch_mixtox_verlauf(rq_ue_beispiel_mvwizr,
  modus = "andauernd",
  plot_zusammenfassung = "mischproben"
)

# Zusammengefasster Verlauf für kurzzeitige Belastungen
plot_misch_mixtox_verlauf(rq_ue_beispiel_mvwizr,
  modus = "kurzzeitig",
  plot_zusammenfassung = "mischproben"
)

# Zusammengefasster Verlauf für Stichproben
plot_misch_mixtox_verlauf(rq_ue_beispiel_mvwizr,
  modus = "kurzzeitig",
  plot_zusammenfassung = "stichproben"
)
}
