% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_misch_ue}
\alias{plot_misch_ue}
\title{GSchV-Überschreitungen in Mischproben plotten}
\usage{
plot_misch_ue(rq_ue_daten, stationscode, jahr = NULL, plot_typ = "andauernd")
}
\arguments{
\item{rq_ue_daten}{Dataframe mit Output der Funktion \code{berechne_rq_ue()}}

\item{stationscode}{Station, für welche der Plot erstellt werden soll.}

\item{jahr}{Jahr, für welches der Überschreitungsplot erstellt werden soll.}

\item{plot_typ}{Legt fest, für welche Art von Überschreitungen eine Auswertung vorgenommen werden soll. Mögliche Werte:
\itemize{
\item \code{"andauernd"}: Berücksichtigt nur Stoffe mit einem spezifischen andauernden Grenzwert in der GSchV. Nur Proben mit einer Dauer von >= 10 Tagen werden verwendet und unter Verwendung des CQK beurteilt.
\item \code{"kurzzeitig"}: Berücksichtigt nur Stoffe mit einem spezifischen akuten Grenzwert in der GSchV. Berücksichtigt alle Proben (ausser Stichproben) und verwendet das AQK zur Beurteilung.
\item \code{"allgemein"}:  Berücksichtigt nur Stoffe, für die kein ökotoxikologisch begründeter Wert in der GSchV vorliegt. Beurteilung mittels allgemeinem Grenzwert (0.1 µg/l)
}}
}
\value{
ggplot2 Plot-Objekt
}
\description{
GSchV-Überschreitungen in Mischproben plotten
}
\examples{
set_ch_locale(lang = "de")

\donttest{
# Andauernde Überschreitungen (und Unterschreitungen) für die Jahre 2019 und 2020 plotten
plot_misch_ue(rq_ue_beispiel_mvwizr, "URT010", jahr = c(2019, 2020))

# Kurzzeitige Überschreitungen für 2019 anzeigen
plot_misch_ue(rq_ue_beispiel_mvwizr, "URT010", plot_typ = "kurzzeitig", jahr = 2019)

# Kurzzeitige Überschreitungen für Station MUS001 für gesamte Zeitdauer in Daten anzeigen
plot_misch_ue(rq_ue_beispiel_mvwizr, "MUS001", plot_typ = "kurzzeitig")
}
}
