% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_misch_ue_qk}
\alias{plot_misch_ue_qk}
\title{Aufsummierte QK-Überschreitungen plotten}
\usage{
plot_misch_ue_qk(
  rq_ue_daten,
  stationscode = NULL,
  jahr = NULL,
  qk = "chronisch",
  detailliert = FALSE
)
}
\arguments{
\item{rq_ue_daten}{Dataframe mit Output der Funktion \code{berechne_rq_ue()}}

\item{stationscode}{Station, für welche der Plot erstellt werden soll. Falls \code{NULL}, wird für alle Stationen geplottet (Facets)}

\item{jahr}{Jahr, für welches der Überschreitungsplot erstellt werden soll. Falls \code{NULL}, werden alle Jahre geplottet.}

\item{qk}{Qualitätskriterium, für welches geplottet werden soll. Mögliche Werte: "chronisch" (Vorgabe) oder "akut".}

\item{detailliert}{Logisch (Vorgabe: \code{FALSE}). Soll der detaillierte VSA-Plot mit Aufspaltung nach Substanzen erstellt werden?}
}
\value{
ggplot2 Plot-Objekt
}
\description{
Aufsummierte QK-Überschreitungen plotten
}
\examples{

# Chronische Übertretungen auswerten
plot_misch_ue_qk(rq_ue_beispiel_mvwizr, qk = "chronisch", detailliert = FALSE)

# Akute Übertretungen auswerten
plot_misch_ue_qk(rq_ue_beispiel_mvwizr, qk = "akut", detailliert = FALSE)

\donttest{
# Chronische Übertretungen detailliert auswerten (VSA-Grafik)
plot_misch_ue_qk(rq_ue_beispiel_mvwizr, qk = "chronisch", detailliert = TRUE)
}

}
