% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_misch_verlauf}
\alias{plot_misch_verlauf}
\title{Verlauf von Mischproben plotten}
\usage{
plot_misch_verlauf(
  mv_daten,
  regulierungen,
  stationscode,
  jahr = NULL,
  id_substanz = NULL,
  zulassungstyp = "[BP]",
  plot_typ = "barplot",
  plot_bg = TRUE,
  plot_parametergruppe = ""
)
}
\arguments{
\item{mv_daten}{Dataframe mit aufbereiteten MV-Daten gemäss Spezifikation}

\item{regulierungen}{Dataframe mit aufbereiteten Regulierungs-Daten gemäss Spezifikation}

\item{stationscode}{Station, für welche der Plot erstellt werden soll}

\item{jahr}{Jahre (numerischer Vektor), für welche ein Verlauf geplottet werden soll. Falls \code{NULL} (Vorgabewert), werden alle verfügbaren Daten geplottet.}

\item{id_substanz}{Legt fest, ob für eine Substanz (Vektorlänge = 1; z.B. \code{id_substanz = 71}), mehrere Substanzen (Vektorlänge > 1; z.B. \code{id_substanz = c(71, 91)}) oder alle Substanzen (Summen) der Verlauf geplottet werden soll. Falls \code{NULL} (Vorgabewert) wird ein Summenplot erstellt}

\item{zulassungstyp}{Filtert nach der Zulassungsart (Spalte \verb{Informationen Recht} in Regulierungstabelle). Wird als regulärer Ausdruck interpretiert. Der Vorgabewert filtert nach Bioziden und Pflanzenschutzmitteln (auch mit ausgelaufener Zulassung). Für die Darstellung aller Substanzen (z.B. für gruppierten Barplot) "Alle" verwenden. Falls nur bestimmte Zulassungsschlüssel verwendet werden sollen, dies als regulären Ausdruck angeben (z.B. \code{zulassungstyp = "B|PX"} für zugelassene Biozide und/oder nicht mehr zugelassene PSM).}

\item{plot_typ}{Plottyp in Abhängigkeit des gewählten Modus bei \code{id_substanz}.
\itemize{
\item Falls \code{length(id_substanz) == 1}, sind die folgenden Werte erlaubt: \code{"barplot"}, \code{"striche"}, \code{"treppen"}. Wenn vorhanden, werden minimale und maximale Bestimmungsgrenzen geplottet
\item Falls \code{length(id_substanz) > 1}, sind die folgenden Werte erlaubt: \code{"striche"}, \code{"treppen"}. Es werden keine Bestimmungsgrenzen geplottet.
\item Falls keine \code{id_substanz} angegeben wird, werden Summenplots gezeichnet. Folgende Werte sind erlaubt: \code{"barplot"}, \code{"kombiniert"}, \code{"treppen"}. \code{"kombiniert"} stellt eine Kombination der anderen beiden Plottypen dar (z.B. für Kontrollzwecke).
}}

\item{plot_bg}{Logisch (Vorgabe: \code{TRUE}). Sollen Bestimmungsgrenzen, falls vorhanden, gezeichnet werden?}

\item{plot_parametergruppe}{Name der Spalte (String), die für das Kategorisieren gruppierter Barplots (Summenplots) verwendet werden soll. Für die Daten des GBL Bern "PARAMETERGRUPPE" verwenden.}
}
\value{
ggplot2 Plot-Objekt
}
\description{
Plottet den Konzentrationsverlauf von Mikroverunreinigungen für eine bestimmte Station und Zeitraum. Dabei werden Stichproben (Proben ohne Enddatum) entfernt.
}
\details{
Es sind in Abhängigkeit des Modus (\code{id_substanz}) verschiedene Plottypen \code{plot_typ} erlaubt:

\itemize{
\item \code{"barplot"}: Monochromer Barplot; entweder für 1 Substanz oder Summenplots. Es werden auch überlappende Messintervalle dargestellt.
\item \code{"kombiniert"}: Eine Kombination von \code{"barplot"} und \code{"treppen"} zur Diagnose/Kontrolle.
\item \code{"striche"}: Horizontale Striche; entweder für 1 oder mehrere Substanzen. Es werden auch überlappende Messintervalle dargestellt.
\item \code{"treppen"}: Treppenplot (ähnlich GeomStep). Kurze Messdauern <10 Tagen werden für diese Darstellung entfernt, um die Wahrscheinlichkeit von Überlappungen zu minimieren. Falls dennoch zeitliche Überlappungen existieren, stoppt die Funktion mit einem Fehler.
\item \code{"barplot_gruppen"}: Nur für Summenplots (\code{id_substanz} = \code{NULL}), wenn \code{plot_parametergruppe} (siehe unten) angegeben ist und diese Spalte in den Daten existiert. Zeigt Barplots gruppiert nach einer bestimmten Kategorisierung (z.B. Aufteilung in Pestizide, Haushaltschemikalien und Arzneimittel). Achtung: Damit Stoffe unabhängig vom Zulassungsstatus gezeigt werden, muss bei \code{zulassungstyp} "Alle" gesetzt sein.
}
}
\examples{
set_ch_locale(lang = "de")

# Summenplot mit Barplot-Darstellung
plot_misch_verlauf(mvdaten_beispiel_mvwizr, regulierungen_mvwizr, "URT010")

# Summenplot mit Treppen-Darstellung
plot_misch_verlauf(mvdaten_beispiel_mvwizr, regulierungen_mvwizr,
  "URT010",
  plot_typ = "treppen"
)

# Summenplot mit Treppen-Darstellung bei überlappenden Intervallen - Fehler!
if (FALSE) {
  plot_misch_verlauf(mvdaten_beispiel_mvwizr, regulierungen_mvwizr,
    "MUS001",
    plot_typ = "treppen"
  )
}

# Verlauf von Einzelsubstanzen mit Barplots und Anzeige der Bestimmungsgrenze:
plot_misch_verlauf(mvdaten_beispiel_mvwizr, regulierungen_mvwizr,
  "URT010",
  id_substanz = 71
)

# Verlauf mehrerer Substanzen mit Strichen:
plot_misch_verlauf(mvdaten_beispiel_mvwizr, regulierungen_mvwizr,
  "URT010",
  plot_typ = "striche", id_substanz = c(71, 91)
)

# Verlauf Summe aller Substanzen (nicht nur Pestizide) als gruppierte Barplots
plot_misch_verlauf(mvdaten_beispiel_mvwizr, regulierungen_mvwizr,
  "URT010",
  plot_typ = "barplot_gruppen", zulassungstyp = "Alle",
  plot_parametergruppe = "PARAMETERGRUPPE"
)
}
