% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_stich_mixtox_zf}
\alias{plot_stich_mixtox_zf}
\title{Zusammenfassung von Mischungstoxizitäten für Stichproben plotten}
\usage{
plot_stich_mixtox_zf(rq_ue_daten, stationscode = NULL, jahr = NULL)
}
\arguments{
\item{rq_ue_daten}{Dataframe mit Output der Funktion \code{berechne_rq_ue()}}

\item{stationscode}{Station, für welche der Plot erstellt werden soll.}

\item{jahr}{Jahr, für welches der Überschreitungsplot erstellt werden soll.}
}
\value{
ggplot2 Plot-Objekt
}
\description{
Hinweis: Hierbei handelt es sich nur um einen Wrapper um \code{plot_misch_mixtox_verlauf} für dieselben Plots aber mit Stichproben.
}
\examples{
plot_stich_mixtox_zf(rq_ue_beispiel_mvwizr)
}
