% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_stich_uebersicht}
\alias{plot_stich_uebersicht}
\title{Stationsübersicht Stichproben}
\usage{
plot_stich_uebersicht(mv_daten, stationscode, jahr = NULL)
}
\arguments{
\item{mv_daten}{Dataframe mit aufbereiteten MV-Daten gemäss Spezifikation}

\item{stationscode}{Station, für welche der Plot erstellt werden soll}

\item{jahr}{Jahre (numerischer Vektor), für welche ein Verlauf geplottet werden soll. Falls \code{NULL} (Vorgabewert), werden alle verfügbaren Daten geplottet.}
}
\value{
ggplot2 Plot-Objekt
}
\description{
Diese Übersichtsfunktion plottet sämtliche Substanzen in allen Stichproben in einem als Raster über die Zeit und benutzt eine logarithmische Farbskala um relevante Substanzen und Veränderungen hervorzuheben.
}
\examples{
# Stichprobenübersichten für zwei Stationen
plot_stich_uebersicht(mvdaten_beispiel_mvwizr, stationscode = "SA51")
plot_stich_uebersicht(mvdaten_beispiel_mvwizr, stationscode = "KI52")
}
