% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_stich_verlauf}
\alias{plot_stich_verlauf}
\title{Stichproben Konzentrationsverlauf}
\usage{
plot_stich_verlauf(
  mv_daten,
  stationscode = NULL,
  jahr = NULL,
  id_substanz = NULL
)
}
\arguments{
\item{mv_daten}{Dataframe mit aufbereiteten MV-Daten gemäss Spezifikation}

\item{stationscode}{Station, für welche der Plot erstellt werden soll}

\item{jahr}{Jahre (numerischer Vektor), für welche ein Verlauf geplottet werden soll. Falls \code{NULL} (Vorgabewert), werden alle verfügbaren Daten geplottet.}

\item{id_substanz}{Substanz(en), die geplottet werden sollen. Falls \code{NULL} werden sämtliche Substanzen geplottet.}
}
\value{
ggplot2 Plot-Objekt
}
\description{
Stichproben Konzentrationsverlauf
}
\examples{
# Plotten aller Substanzen für alle Stationen - sehr unübersichtlich
pobj <- plot_stich_verlauf(mvdaten_beispiel_mvwizr)

# Allerdings eignet sich die Darstellung für die interaktive Analyse mit Plotly.
if (rlang::is_installed("plotly")) plotly::ggplotly(pobj)
}
