% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regulierungen_mvwizr}
\alias{regulierungen_mvwizr}
\title{Regulierungen}
\format{
\subsection{\code{regulierungen_mvwizr}}{

Tabelle mit 4 Variablen und knapp 3000 Einträgen. Im Folgenden sind die Variablen mit ihren Datentypen beschrieben:
\describe{
\item{ID_Substanz}{VSA-ID der Substanz (integer)}
\item{Name_reg}{Bezeichnung, die vom VSA verwendet wird (character)}
\item{Informationen Recht}{Zeichenkette mit codierten Regulierungsinformationen (character)}
\item{GSCHV}{Gibt an, ob und wie die Substanz in der GSCHV reguliert ist. 1 = Spezifischer Wert in Anh. 2; 2 = allgemeiner Grenzwert von 0.1 µg/l (integer)}
}
}
}
\source{
VSA
}
\usage{
regulierungen_mvwizr
}
\description{
Dieser Datensatz beinhaltet Informationen zur Regulierung von Mikroverunreinigungen.
}
\keyword{datasets}
