% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rq_ue_beispiel_mvwizr}
\alias{rq_ue_beispiel_mvwizr}
\title{RQ und Ue aus MV-Beispieldaten}
\format{
\subsection{\code{rq_ue_beispiel_mvwizr}}{

Aufgrund des Joins der Eingangsdaten besteht dieser Datensatz aus 52 (oder mehr) Variablen, wobei viele davon bereits in der Hilfe zu den Eingangsdaten beschrieben sind. Im Folgenden werden daher nur die zusätzlich berechneten Variablen aufgeführt:
\describe{
\item{RQ_CQK}{Risikoquotient für CQK allgemein (numeric)}
\item{RQ_AQK}{Risikoquotient für AQK allgemein (numeric)}
\item{RQ_CQK_P}{Risikoquotient für CQK für Pflanzen (numeric)}
\item{RQ_CQK_I}{Risikoquotient für CQK für Invertebraten (numeric)}
\item{RQ_CQK_V}{Risikoquotient für CQK für Vertebraten (numeric)}
\item{RQ_AQK_P}{Risikoquotient für AQK für Pflanzen (numeric)}
\item{RQ_AQK_I}{Risikoquotient für AQK für Invertebraten (numeric)}
\item{RQ_AQK_V}{Risikoquotient für AQK für Vertebraten (numeric)}
\item{Beurteilung_CQK}{Beurteilung für CQK gem. MSK in 5 Stufen (factor)}
\item{Beurteilung_AQK}{Beurteilung für CQK gem. MSK in 5 Stufen (factor)}
\item{Ue_anhaltend}{Übertretung Grenzwert anhaltende Belastung? (logical)}
\item{Ue_kurzzeitig}{Übertretung Grenzwert kurzzeitige Belastung? (logical)}
\item{Ue_spezifisch}{Übertretung Grenzwert spezifische Belastungswerte (d.h. kurzzeitig oder spezifisch)? (logical)}
\item{Ue_generisch}{Übertretung generischer Grenzwert? (logical)}
\item{Ue_AQK}{Übertretung akutes QK (Robustheit 1,2)? (logical)}
\item{Ue_CQK}{Übertretung chronisches QK (Robustheit 1,2)? (logical)}
}
}
}
\source{
GBL/VSA
}
\usage{
rq_ue_beispiel_mvwizr
}
\description{
Dieser Datensatz beinhaltet Risikoquotienten und GSchV-Überschreitungen zu den MV-Beispieldaten und wurde aus \code{mvdaten_beispiel_mvwizr} berechnet.
}
\keyword{datasets}
