% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{schreibe_nawa_import_manifest_template}
\alias{schreibe_nawa_import_manifest_template}
\title{Import Manifest Vorlage für NAWA-MV-Daten schreiben}
\usage{
schreibe_nawa_import_manifest_template(
  import_manifest_file,
  mv_data_pfad = NULL
)
}
\arguments{
\item{import_manifest_file}{Pfad zur Import-Manifest-Datei, die geschrieben werden soll. Diese Datei wird im Excel-Format (.xlsx) geschrieben.}

\item{mv_data_pfad}{Pfad zu den MV-Daten, die im NAWA-Format vorliegen. Falls angegeben, werden alle Dateien im Verzeichnis, die dem NAWA-Format entsprechen, in das Manifest aufgenommen. Falls \code{NULL}, wird ein leeres Manifest geschrieben.}
}
\value{
Gibt das Manifest-Template als tibble zurück.
}
\description{
Import Manifest Vorlage für NAWA-MV-Daten schreiben
}
\examples{

if (FALSE) {
  template_path <- "nawa_import_manifest_template.xlsx"
  data_dir <- "/Users/username/Documents/NAWA_MV_Daten"
  schreibe_nawa_import_manifest_template(template_path, data_dir)
}

}
