% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_ch_locale}
\alias{set_ch_locale}
\title{Set CH locale globally}
\usage{
set_ch_locale(lang = c("de", "fr"))
}
\arguments{
\item{lang}{Sprache für locale. Entweder "de" oder "fr"}
}
\value{
TRUE
}
\description{
Diese Funktion setzt (plattformabhängig) die korrekte locale (Spracheinstellung) für die Erzeugung der Plots.
}
