% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvwizr-package.R
\docType{package}
\name{mvwizr-package}
\alias{mvwizr}
\alias{mvwizr-package}
\title{mvwizr: Mikroverunreinigungen in Schweizer Gewässern darstellen}
\description{
Tools um Mikroverunreinigungsdaten aus Schweizer Gewässern gemäss den Bestimmungen der Gewässerschutzverordnung (GSchV) darzustellen.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ror-at-ebp.github.io/mvwizr/}
}

}
\author{
\strong{Maintainer}: Remo Röthlin \email{remo.roethlin@ebp.ch}

Other contributors:
\itemize{
  \item Veronica Bozzini \email{veronica.bozzini@ebp.ch} [contributor]
  \item Irene Wittmer \email{irene.wittmer@be.ch} [contributor]
  \item Kanton Bern, AWA [copyright holder, funder]
}

}
\keyword{internal}
