% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_misch_oekotox_uebersicht}
\alias{plot_misch_oekotox_uebersicht}
\title{Stationsübersicht Ökotoxbeurteilung plotten}
\usage{
plot_misch_oekotox_uebersicht(
  rq_ue_daten,
  stationscode,
  jahr,
  modus = "andauernd"
)
}
\arguments{
\item{rq_ue_daten}{Dataframe mit Output der Funktion \code{berechne_rq_ue()}}

\item{stationscode}{Station, für welche der Plot erstellt werden soll}

\item{jahr}{Jahr, für welches der Plot erstellt werden soll}

\item{modus}{\describe{
\item{"andauernd"}{Verwendung des CQK für die Beurteilung & Proben mit Dauer >= 10 Tage}
\item{"kurzzeitig"}{Verwendung des AQK für die Beurteilung}
}}
}
\value{
ggplot2 Plot-Objekt
}
\description{
Stationsübersicht Ökotoxbeurteilung plotten
}
\examples{

plot_misch_oekotox_uebersicht(rq_ue_beispiel_mvwizr, "URT010", 2020)

plot_misch_oekotox_uebersicht(rq_ue_beispiel_mvwizr, "MUS001", 2020, modus = "kurzzeitig")

plot_misch_oekotox_uebersicht(rq_ue_beispiel_mvwizr, "MUS001", 2020)
}
