% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_misch_ue_qk}
\alias{plot_misch_ue_qk}
\title{Aufsummierte QK-Überschreitungen plotten}
\usage{
plot_misch_ue_qk(
  rq_ue_daten,
  stationscode = NULL,
  jahr = NULL,
  qk = c("chronisch", "akut"),
  detailliert = FALSE,
  pattern_key_scale = 1
)
}
\arguments{
\item{rq_ue_daten}{Dataframe mit Output der Funktion \code{berechne_rq_ue()}}

\item{stationscode}{Station, für welche der Plot erstellt werden soll.}

\item{jahr}{Jahr, für welches der Überschreitungsplot erstellt werden soll.}

\item{qk}{Qualitätskriterium, für welches geplottet werden soll. Mögliche Werte: "chronisch" (Vorgabe) oder "akut".}

\item{detailliert}{Logisch (Vorgabe: \code{FALSE}). Soll der detaillierte VSA-Plot mit Aufspaltung nach Substanzen erstellt werden?}

\item{pattern_key_scale}{Skalierungsfaktor (Vorgabe: 1), der angibt, wie das Muster in den Legendeneinträgen skaliert werden soll (nur für \code{detailliert = TRUE}). Diesen Wert reduzieren, falls ein Muster im Plot gezeigt wird, aber nicht in der Legende.}
}
\value{
ggplot2 Plot-Objekt
}
\description{
Zeigt die Überschreitung chronischer oder akuter Qualitätskriterien pro Jahr und Station, wobei sichtbar wird, ob es viele relevante Überschreitungen gibt von Qualitätskriterien, die nicht in der GSchV aufgenommen sind als spezifischer Grenzwert.
}
\examples{
# Chronische Übertretungen auswerten
plot_misch_ue_qk(rq_ue_beispiel_mvwizr, qk = "chronisch", detailliert = FALSE)

# Akute Übertretungen auswerten
plot_misch_ue_qk(rq_ue_beispiel_mvwizr, qk = "akut", detailliert = FALSE)

\donttest{
# Chronische Übertretungen detailliert auswerten (VSA-Grafik)
# Wird nicht automatisch getestet, weil Run zu lange dauert
plot_misch_ue_qk(rq_ue_beispiel_mvwizr, qk = "chronisch", detailliert = TRUE)
}

}
