% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{schreibe_nawa_import_manifest_template}
\alias{schreibe_nawa_import_manifest_template}
\title{Schreibe Import Manifest Vorlage für NAWA-MV-Daten}
\usage{
schreibe_nawa_import_manifest_template(
  import_manifest_file,
  mv_datei_pfade = NULL
)
}
\arguments{
\item{import_manifest_file}{Pfad zur Import-Manifest-Datei, die geschrieben werden soll. Diese Datei wird im Excel-Format (.xlsx) geschrieben.}

\item{mv_datei_pfade}{Dateipfade zu den MV-Daten, die im NAWA-Format vorliegen. Falls angegeben, werden alle Dateien in das Manifest aufgenommen. Falls \code{NULL}, wird ein leeres Manifest geschrieben.}
}
\value{
Gibt das Manifest-Template als tibble zurück.
}
\description{
Schreibe Import Manifest Vorlage für NAWA-MV-Daten
}
\examples{

nawa_mv_pfade <- system.file("extdata", c(
  "NAWA_Bsp_1.xlsx",
  "NAWA_Bsp_2.xlsx", "NAWA_Bsp_3.csv", "NAWA_Bsp_4.xlsx",
  "NAWA_Bsp_5.csv"
), package = "mvwizr")

schreibe_nawa_import_manifest_template(tempfile(fileext = ".xlsx"),
  mv_datei_pfade = nawa_mv_pfade
)

}
